RAG_education/backend/src/config.py
"""設定定数"""
import os
from pathlib import Path
from dotenv import load_dotenv

load_dotenv()

BASE_DIR = Path(__file__).resolve().parent.parent
DATA_DIR = BASE_DIR / "data"
CHROMA_DIR = BASE_DIR / "chroma_db"

EMBEDDING_MODEL = "intfloat/multilingual-e5-base"
CHROMA_COLLECTION = "rag_education"

PORT = int(os.getenv("PORT", "7000"))
CORS_ORIGINS = os.getenv("CORS_ORIGINS", "http://localhost:3000").split(",")
