RAG_education/backend/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.20.0
python-dotenv>=1.0.0
langchain>=0.2.0
langchain-community>=0.2.0
langchain-google-genai>=4.0.0
chromadb>=0.4.0
sentence-transformers>=2.2.0
pydantic>=2.0.0
